+++ src/main/java/net/byteseek/incubator/matcher/multisequence/MultiSequenceReverseMatcher.java
@@ -225,7 +225,7 @@ public final class MultiSequenceReverseMatcher implements MultiSequenceMatcher {
 
 
     @Override
-    public long matches(WindowReader reader, long matchPosition, List<MatchResult> results) throws IOException {
+    public int matches(WindowReader reader, long matchPosition, List<MatchResult> results) throws IOException {
         return 0; //TODO: implement.
     }
 
@@ -238,7 +238,7 @@ public final class MultiSequenceReverseMatcher implements MultiSequenceMatcher {
     }
 
     @Override
-    public long matches(byte[] bytes, int matchPosition, List<MatchResult> results) {
+    public int matches(byte[] bytes, int matchPosition, List<MatchResult> results) {
         return 0; //TODO: implement.
     }
